refactor: 서비스 클래스명, 테스트 클래스명 변경
